Make homebrew install more smart about install vs updating
rejigger script to test whether apps are already installed. Only try to install packages or casks if they are not installed. Test to see if dupes and caskroom taps are tapped.

Should fix issue with caskroom install
Should now be able to run script safely regardless of current computer state